Changes conerting the project into a Spring Data JPA
