+++ src/main/java/com/mandu/productManagement/ProductSprintApplication.java
@@ -1,0 +1,18 @@
+package com.mandu.productManagement;
+
+import org.springframework.context.ApplicationContext;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+//@SpringBootApplication
+public class ProductSprintApplication {
+//    public static void main(String[] args) {
+//        //create application context
+//        ApplicationContext context = SpringApplication.run(ProductSprintApplication.class,args);
+
+
+
+    }
+
+
+
