+++ src/main/resources/application.properties
@@ -1,1 +1,5 @@
 spring.application.name=productManagementProj
+
+spring.datasource.url=jdbc:mysql://localhost:3306/springjdbc
+spring.datasource.username=root
+spring.datasource.password=Mandu
